# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Open Card Control Plane API
  version: 1.1
  description: >-
    REST JSON and persistent SSE control-plane endpoints. Product state remains owned by the
    control plane; provider and agent endpoints are versioned contracts. API clients may send
    an exact Open-Card-API-Version header or an Open-Card-API-Min-Version and
    Open-Card-API-Max-Version range. All version headers are optional for backwards
    compatibility: clients that omit them negotiate the legacy 1.0 contract. Requests with
    no common supported version receive 426 Upgrade Required.
servers:
  - url: /
paths:
...
  /api/v1/applications:
    parameters:
      - $ref: '#/components/parameters/APIVersion'
      - $ref: '#/components/parameters/APIMinVersion'
      - $ref: '#/components/parameters/APIMaxVersion'
    get:
      operationId: listApplications
      responses:
        '200':
          description: Applications
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
        '426':
          $ref: '#/components/responses/APIUpgradeRequired'
    post:
      operationId: createApplication
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Created
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApplicationResponse'
        '400':
          description: Invalid request
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
        '409':
          description: Idempotency key was reused with different input
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
        '426':
          $ref: '#/components/responses/APIUpgradeRequired'
...
  /api/v1/applications/{applicationId}/usage:
    parameters:
      - $ref: '#/components/parameters/ApplicationID'
      - $ref: '#/components/parameters/APIVersion'
      - $ref: '#/components/parameters/APIMinVersion'
      - $ref: '#/components/parameters/APIMaxVersion'
    get:
      operationId: getApplicationUsage
      description: >-
        Reads local aggregate usage facts from independent runtime observations. The interval is
        UTC and half-open [from,to). Ordinary mode is a comprehensible summary; operations mode
        includes service/release/time trends, actual values, configured limits, and evidence-linked
        anomalies. No price, plan, order, payment, balance, invoice, or regional claim is returned.
      parameters:
        - {name: mode, in: query, schema: {type: string, enum: [normal, operations], default: normal}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
        - {name: service, in: query, schema: {type: string, maxLength: 128}}
        - {name: release, in: query, schema: {type: string}}
        - {name: Open-Card-Role, in: header, required: false, schema: {type: string, enum: [operator]}}
      responses:
        '200':
          description: Application-scoped M5 usage facts. AI remains disabled.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApplicationUsageFact'}
        '400': {description: Invalid filter or a window longer than 31 days}
        '403': {description: Operations detail requires operator role}
        '404': {description: Application not found}
        '503': {description: Local meter or usage persistence unavailable}
...
components:
  parameters:
    APIVersion:
      name: Open-Card-API-Version
      in: header
      required: false
      description: >-
        Request an exact API version. This header is mutually exclusive with the minimum and
        maximum version range headers. If all version headers are omitted, API version 1.0 is
        negotiated for legacy clients.
      schema:
        type: string
        pattern: '^v?[0-9]+(\.[0-9]+)?$'
    APIMinVersion:
      name: Open-Card-API-Min-Version
      in: header
      required: false
      description: Lowest API version accepted by the client. Use with Open-Card-API-Max-Version.
      schema:
        type: string
        pattern: '^v?[0-9]+(\.[0-9]+)?$'
    APIMaxVersion:
      name: Open-Card-API-Max-Version
      in: header
      required: false
      description: Highest API version accepted by the client. Use with Open-Card-API-Min-Version.
      schema:
        type: string
        pattern: '^v?[0-9]+(\.[0-9]+)?$'
    ApplicationID:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Stable retry key. Reuse with different input is rejected.
      schema:
        type: string
        minLength: 1
...
  headers:
    APINegotiatedVersion:
      description: API version selected for this request and response.
      schema:
        type: string
        pattern: '^[0-9]+\.[0-9]+$'
  responses:
    APIUpgradeRequired:
      description: No common API version is available for the requested range.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  schemas:
    APIError:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
        message:
          type: string
    Application:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ApplicationList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Application'
    CreateApplicationResponse:
      type: object
      required: [application, operation_id]
      properties:
        application:
          $ref: '#/components/schemas/Application'
        operation_id:
          type: string
    CreateApplicationRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
...
    UsageMeasurement:
      type: object
      additionalProperties: false
      required: [actual, limit, unit]
      properties:
        actual: {type: number, minimum: 0}
        limit: {type: number, minimum: 0, description: Zero means no configured limit fact is available.}
        unit: {type: string, enum: [mCPU, B]}
    UsageResourceMeasurements:
      type: object
      additionalProperties: false
      required: [cpu, memory, disk, network]
      properties:
        cpu: {$ref: '#/components/schemas/UsageMeasurement'}
        memory: {$ref: '#/components/schemas/UsageMeasurement'}
        disk: {$ref: '#/components/schemas/UsageMeasurement'}
        network: {$ref: '#/components/schemas/UsageMeasurement'}
    UsageTrendPoint:
      allOf:
        - {$ref: '#/components/schemas/UsageResourceMeasurements'}
        - type: object
          required: [observed_at]
          properties:
            observed_at: {type: string, format: date-time}
    UsageAnomaly:
      type: object
      additionalProperties: false
      required: [id, observed_at, severity, summary]
      properties:
        id: {type: string}
        observed_at: {type: string, format: date-time}
        severity: {type: string, enum: [info, warning, critical]}
        summary: {type: string}
        related_service_id: {type: string}
        related_release_id: {type: string}
    UsageServiceFact:
      type: object
      additionalProperties: false
      required: [id, name, release_id, runtime, started_at, actual, average, peak, configured, trend, anomalies]
      properties:
        id: {type: string}
        name: {type: string}
        release_id: {type: string}
        runtime: {type: string}
        started_at: {type: string, format: date-time}
        actual: {$ref: '#/components/schemas/UsageResourceMeasurements'}
        average: {$ref: '#/components/schemas/UsageResourceMeasurements'}
        peak: {$ref: '#/components/schemas/UsageResourceMeasurements'}
        configured: {$ref: '#/components/schemas/UsageResourceMeasurements'}
        trend: {type: array, items: {$ref: '#/components/schemas/UsageTrendPoint'}}
        anomalies: {type: array, items: {$ref: '#/components/schemas/UsageAnomaly'}}
    ApplicationUsageFact:
      type: object
      additionalProperties: false
      required: [version, observed_at, application_id, application_name, services, anomalies, ai]
      properties:
        version: {type: string, pattern: '^usage-v1:[0-9a-f]{24}$'}
        observed_at: {type: string, format: date-time}
        application_id: {type: string}
        application_name: {type: string}
        services: {type: array, items: {$ref: '#/components/schemas/UsageServiceFact'}}
        anomalies: {type: array, items: {$ref: '#/components/schemas/UsageAnomaly'}}
        ai: {type: string, enum: [disabled]}
...